backend/pyproject.toml
[project]
name = "portfolio-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "google-genai>=1.0",
  "chromadb>=0.5",
  "sentence-transformers>=3.0",
  "rank-bm25>=0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
  "ruff>=0.5",
]

[tool.ruff]
line-length = 100

